requirements.txt
langchain-community~=0.3.21
langchain-huggingface
langchain-openai
aiofiles
aiogram
pypdf2
python-docx
sentence-transformers
python-dotenv~=1.0.1
python-jose
passlib>=1.7.4
bcrypt==3.2.0
SQLAlchemy~=2.0.40
python-multipart
uvicorn
pydantic~=2.10.6
pydantic_core~=2.27.2